Resource usage improvements
Writing the output.csv could consume a large amount of CPU when the generated report was large.  The code original wrote the variable containing the report in one line.  Now the code iterates through the variable in chucks of a few million bytes, writing the chucks and pausing for a second. This drastically reduced CPU usage.